+++ source/ProcessManager.Orchestrations.Abstractions/Processes/BRS_021/ForwardMeteredData/V1/Model/Brs_021_ForwardMeteredData_V1.cs
@@ -14,7 +14,7 @@
 
 using Energinet.DataHub.ProcessManager.Abstractions.Api.Model.OrchestrationDescription;
 
-namespace Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_021.MeteredDataForwarding.V1.Model;
+namespace Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_021.ForwardMeteredData.V1.Model;
 
-public record Brs_021_MeteredDataForwarding_V1()
-    : OrchestrationDescriptionUniqueNameDto("BRS_021_MeteredDataForwarding", 1);
+public record Brs_021_ForwardedMeteredData_V1()
+    : OrchestrationDescriptionUniqueNameDto("BRS_021_ForwardMeteredData", 1);