+++ docs/ProcessManager.Orchestrations.Abstractions/ReleaseNotes/ReleaseNotes.md
@@ -2,8 +2,8 @@
 
 ## Version 0.2.0
 
-- Implemented `Brs_021_MeteredDataForwarding_V1`
-- Implemented `StartMeteredDataForwardingCommandV1`
+- Implemented `Brs_021_ForwardMeteredData_V1`
+- Implemented `StartForwardMeteredDataCommandV1`
 
 ## Version 0.1.0